feat: animate menu item selection
