+++ lib/widgets/menu_item.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:pass_emploi_app/widgets/selected_menu_item.dart';
-import 'package:pass_emploi_app/widgets/unselected_menu_item.dart';
+
+import 'default_menu_item.dart';
 
 class MenuItem extends BottomNavigationBarItem {
   final String drawableRes;
@@ -9,8 +9,8 @@ class MenuItem extends BottomNavigationBarItem {
 
   MenuItem({required this.drawableRes, required this.label, this.withBadge = false})
       : super(
-          icon: UnselectedMenuItem(drawableRes: drawableRes, label: label, withBadge: withBadge),
-          activeIcon: SelectedMenuItem(drawableRes: drawableRes, label: label),
+          icon: DefaultMenuItem(drawableRes: drawableRes, label: label, isActive: false, withBadge: withBadge),
+          activeIcon: DefaultMenuItem(drawableRes: drawableRes, label: label, isActive: true, withBadge: false),
           label: label,
         );
 }
